converted more test to run with 'waitForAsyncResult' and moved that function to tests/support/util.ts
